Changed the way the acquirer url is set and used to support ABN AMRO.
I also added the option of setting the acquirer url(s) by setting a single attribute 'acquirer'.
See the README for more info on this.

Note: remote tests for ABN AMRO are not supported the way they are with Rabobank and ING.
You have to do the tests manually, one by one.